+++ Makefile-rpm-ostree.am
@@ -48,8 +48,8 @@ librpmostreeinternals_la_SOURCES = \
 	src/app/rpmostree-db-builtin-diff.cxx \
 	src/app/rpmostree-db-builtin-list.cxx \
 	src/app/rpmostree-db-builtin-version.cxx \
-	src/app/rpmostree-dbus-helpers.cxx \
-	src/app/rpmostree-dbus-helpers.h \
+	src/app/rpmostree-clientlib.cxx \
+	src/app/rpmostree-clientlib.h \
 	src/app/rpmostree-override-builtins.h \
 	src/app/rpmostree-override-builtins.cxx \
 	src/app/rpmostree-libbuiltin.cxx \
