+++ src/app/rpmostree-builtin-applylive.cxx
@@ -25,7 +25,7 @@
 
 #include "rpmostree-ex-builtins.h"
 #include "rpmostree-libbuiltin.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 #include "rpmostree-cxxrs.h"
 
 #include <libglnx.h>
