+++ src/app/rpmostree-builtin-initramfs.cxx
@@ -25,7 +25,7 @@
 
 #include "rpmostree-builtins.h"
 #include "rpmostree-libbuiltin.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 
 #include <libglnx.h>
