+++ src/app/rpmostree-builtin-rebase.cxx
@@ -27,7 +27,7 @@
 #include "rpmostree-util.h"
 #include "rpmostree-core.h"
 #include "rpmostree-libbuiltin.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 
 #include <libglnx.h>
