+++ src/app/rpmostree-builtin-reset.cxx
@@ -26,7 +26,7 @@
 #include "rpmostree-ex-builtins.h"
 #include "rpmostree-util.h"
 #include "rpmostree-libbuiltin.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 
 #include <libglnx.h>
