+++ src/app/rpmostree-builtin-upgrade.cxx
@@ -27,7 +27,7 @@
 #include "rpmostree-builtins.h"
 #include "rpmostree-libbuiltin.h"
 #include "rpmostree-rpm-util.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 
 #include <libglnx.h>
