+++ src/app/rpmostree-clientlib.cxx
@@ -32,7 +32,7 @@
 #include <libglnx.h>
 
 #include "rpmostree-types.h"
-#include "rpmostree-dbus-helpers.h"
+#include "rpmostree-clientlib.h"
 #include "rpmostree-builtins.h"
 #include "rpmostree-libbuiltin.h"
 #include "rpmostree-util.h"
